working bot with 4 buttons

+++ app/schemas/nutrition.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 class NutritionSchema(BaseModel):
     user_telegram_id: int
     calories: int
-    water: float
+    water: int
 
     class Config:
         orm_mode = True
